requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
python-multipart==0.0.6
pytesseract==0.3.10
pdfplumber==0.8.0
PyMuPDF==1.22.5
Pillow==10.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.34.0
openai==1.29.0
aiofiles==23.1.0
requests==2.31.0
typing_extensions
loguru
